#168 fix: 밈테스트 결과 요금제 추천 UI 연동 중 변화된 프로젝트 구조에 최적화
